requirements.txt
ase>=3.22.1
numpy>=1.21.0
pyyaml>=5.4.1
matplotlib>=3.3.0
pytest>=6.0.0
